+++ R/messages.R
@@ -13,32 +13,8 @@ messages$errorEnumPutListMultipleKeys <- function() {
   paste0("Trying to put multiple keys, but only one key is allowed!")
 }
 
-messages$errorFileNotFound <- function(filePath, optionalMessage = NULL) {
-  paste0("File '", filePath, "' could not be found!")
-}
-
-messages$errorEsqlabsRSettingNotFound <- function(settingName) {
-  paste0("No global setting with the name '", settingName, "' exists. Available global settings are:\n", paste0(names(esqlabsEnv), collapse = ", "))
-}
-
 messages$errorDistributionNotSupported <- function(string) {
   paste0("The distribution '", string, "' is not supported. Supported distributions are listed in `Distributions`.")
-}
-
-messages$errorOutputPathNotFound <- function(string) {
-  paste0("The output with the path '", string, "' is not found.")
-}
-
-messages$warningLabelNotInDataMapping <- function(string) {
-  paste0("No xy-series with label ", string, " exists in the DataMapping. Nothing to remove")
-}
-
-messages$errorValuesAreNotPositive <- function(values, optionalMessage = NULL) {
-  paste("All values must be positive or 0, but they are not! Values are: ", paste(as.character(values), collapse = ", "), optionalMessage)
-}
-
-messages$errorWrongLength <- function(object, length, optionalMessage = NULL) {
-  paste("Object `", object, "` must be of length ", length, " but it is not!")
 }
 
 messages$errorWrongPopulationName <- function(populationName) {
@@ -53,16 +29,12 @@ messages$warningValueWithinThresholdNotExisting <- function(value, threshold, op
   paste("value `", value, "` not found in the array within the absolute threshold of  ", threshold, optionalMessage)
 }
 
-messages$errorMultipleMetaDataEntries <- function(optionalMessage = NULL) {
-  paste("Can only set a single meta data entry at once", optionalMessage)
-}
-
 messages$errorInvalidMeanMethod <- function() {
-  paste("Invalid value for argument `method`, supported values are `arithmetic` or `geometric`")
+  "Invalid value for argument `method`, supported values are `arithmetic` or `geometric`"
 }
 
 messages$errorOutputMolWeightNeeded <- function() {
-  paste("`outputMolWeight` can not be `NULL` when data sets have different molWeights")
+  "`outputMolWeight` can not be `NULL` when data sets have different molWeights"
 }
 
 messages$wrongParametersStructure <- function(argumentName) {
@@ -74,4 +46,20 @@ messages$wrongParametersStructure <- function(argumentName) {
 
 messages$valueShouldNotBeNegative <- function(parameterName, value) {
   paste0(parameterName, " must be a positive numerical value, but the value is ", value)
+}
+
+messages$nrOfColorsShouldBePositive <- function(nrOfColors) {
+  paste0("nrOfColors must be positive, value ", nrOfColors, " is not valid!")
+}
+
+messages$wrongSimulationType <- function() {
+  "Wrong value for 'simulationType'! Accepted values are 'Individual and 'Population'"
+}
+
+messages$noPKDataToWrite <- function() {
+  "`pkDataFilePath` argument is specified, but there is no PK parameters data to write to spreadsheets."
+}
+
+messages$cannotGetMoleculeFromQuantity <- function(quantityPath, optionalMessage = NULL) {
+  paste0("Could not retrieve molecule name for the quantity with the path '", quantityPath, "'. ", optionalMessage)
 }