+++ R/utilities-figures.R
@@ -37,7 +37,7 @@ esqLABS_colors <- function(nrOfColors) {
   deltaV_r_g <- max(esqRed_hsv[3], esqGreen_hsv[3]) - min(esqRed_hsv[3], esqGreen_hsv[3])
 
   if (nrOfColors < 0) {
-    stop("nrOfColors must be positive, value ", nrOfColors, " is not valid!")
+    stop(messages$nrOfColorsShouldBePositive(nrOfColors))
   }
   if (nrOfColors == 0) {
     return(c())
@@ -97,51 +97,6 @@ esqLABS_colors <- function(nrOfColors) {
   return(palette)
 }
 
-#' Add a label to a figure. Taken from <https://waterprogramming.wordpress.com/2015/12/02/easy-labels-for-multi-panel-plots-in-r/>
-#'
-#' @param label Label that is to be drawn
-#' @param location Location of the label. Allowed entries are 'topleft',
-#'   'topcenter', 'topright', 'bottomleft', 'bottomright', 'bottomcenter'
-#' @param offset Coordinates of the offset of the label
-#' @keywords internal
-figureAddLabel <- function(label, location = "topleft",
-                           offset = c(0, 0)) {
-  coords <- switch(location,
-    topleft = c(0.015, 0.98),
-    topcenter = c(0.5525, 0.98),
-    topright = c(0.985, 0.98),
-    bottomleft = c(0.015, 0.02),
-    bottomcenter = c(0.5525, 0.02),
-    bottomright = c(0.985, 0.02),
-    c(0.015, 0.98)
-  )
-
-  this.x <- grconvertX(coords[1] + offset[1], from = "nfc", to = "user")
-  this.y <- grconvertY(coords[2] + offset[2], from = "nfc", to = "user")
-  text(labels = label[1], x = this.x, y = this.y, xpd = T)
-}
-
-#' Add legend to a plot
-#'
-#' @inheritParams graphics::legend
-#' @param ... Additional arguments to be passed to `graphics::legend`.
-#' @keywords internal
-.figureAddLegend <- function(x, legend, col, pch, lty, ...) {
-  # Legend does not ignore ellipsis arguments that are not supported, so they
-  # must be removed from the arguments list
-  supportedArgs <- names(formals(graphics::legend))
-  args <- list(...)
-  args <- args[which(names(args) %in% supportedArgs)]
-  args$x <- x
-  args$legend <- legend
-  args$col <- col
-  args$pch <- pch
-  args$lty <- lty
-
-  # Using do.call use arguments combined from ellipsis
-  do.call(graphics::legend, args)
-}
-
 #' Returns the HSV values for a given R color name
 #'
 #' @param color vector of any of the three kinds of R color specifications,
@@ -153,7 +108,7 @@ figureAddLabel <- function(label, location = "topleft",
 #'   indicate hue, saturation and value and are named "h", "s", and "v"
 #'   accordingly.
 #' @export
-#' @import ospsuite
+#' @import ospsuite ospsuite.utils
 #'
 #' @examples
 #' col2hsv("yellow")
@@ -164,125 +119,105 @@ col2hsv <- function(color) {
   return(rgb2hsv(rgb))
 }
 
-#' A function to add error bars on the chart.
-#' @description Taken from <https://www.r-graph-gallery.com/4-barplot-with-error-bar.html>
-#' @import ospsuite
-#'
-#' @param x Numerical array of x-values
-#' @param y Numerical array of y-values
-#' @param upper Numerical array of upper y-error values
-#' @param lower Numerical array of lower y-error values. Optional.
-#' If not specified, same values as for `upper` are used
-#' @param length Numerical value specifying the width of the error bars.
-#'   Optional. Default is 0.1.
-#' @param axis Dimension to which the error bars are added. If "y" (default),
-#'   vertical error bars are drawn. If "x", horizontal error bars are drawn
-#' @param ... Graphical parameters (see [par()])
-#'
-#' @keywords internal
-plotErrorBars <- function(x,
-                          y,
-                          upper,
-                          lower = upper,
-                          length = par()$cin[[1]] / 2,
-                          axis = "y",
-                          ...) {
-  validateIsNumeric(c(x, y))
-  validateIsSameLength(x, y, upper, lower)
-  if (axis == "y") {
-    arrows(x, y + upper, x, y - lower, angle = 90, code = 3, length = length, ...)
-  }
-  if (axis == "x") {
-    arrows(x + upper, y, x - lower, y, angle = 90, code = 3, length = length, ...)
-  }
-}
-
-#' Possible entries for the `outputDevice` field of a `PlotConfiguration` object
+#' Possible entries for the `outputDevice` field of a `ProjectConfiguration` object
 #' @export
 GraphicsDevices <- enum(list("PNG"))
 
-#' Possible entries for the `plotType` field of a `DataMapping` object
+
+#' @title Create an instance of `DefaultPlotConfiguration` R6 class
+#' @rdname createEsqlabsPlotConfiguration
 #'
-#' @details "IndividualProfile" - simulated results are plotted as time-values
-#'   series with points connected by lines, with each individual results potted
-#'   separately "PopulationQuantiles" - simulated results for a population are
-#'   aggregated as median, 95th an 5th percentiles, with median plotted as a
-#'   line an upper/lower percentiles plotted as shaded areas
-#'   "PredictedVsObserved" -  predicted-versus-observed goodness of fit plot
+#' @description
+#'
+#' An instance of `DefaultPlotConfiguration` R6 class is needed for creating
+#' visualizations with the `{ospsuite}` package.
+#'
+#' The default attributes of the class are chosen to reflect the corporate
+#' standards adopted by esqLABS GmbH.
+#'
+#' @return An instance of `DefaultPlotConfiguration` R6 class.
+#'
+#' @examples
+#' createEsqlabsPlotConfiguration()
+#'
+#' @family create-plotting-configurations
+#'
 #' @export
-PlotTypes <- enum(list(
-  "IndividualProfile",
-  "PopulationQuantiles",
-  "PredictedVsObserved",
-  "BoxPlot"
-))
+createEsqlabsPlotConfiguration <- function() {
+  defaultPlotConfiguration <- ospsuite::DefaultPlotConfiguration$new()
 
-#' Open an output device.
-#'
-#' @param plotConfiguration An object of type `PlotConfiguration`
-#' @param width Width of the output figure
-#' @param height Height of the output figure
-#'
-#' @details If the output of the plot is directed to a file, open the
-#'   connection. A list of supported outputs is provided in
-#'   `GraphicsDevices`-enum. If the provided output defined in
-#'   PlotConfiguration$outputDevice is not supported or the value is `NULL`,
-#'   output is directed to the default plot frame.
-#'
-#' @keywords internal
-openOuptutDevice <- function(plotConfiguration, width, height) {
-  if (is.null(plotConfiguration$outputDevice)) {
-    return()
-  }
-  if (plotConfiguration$outputDevice == GraphicsDevices$PNG) {
-    png(
-      filename = file.path(
-        plotConfiguration$outputFolder,
-        paste0(plotConfiguration$outputName, ".png")
-      ),
-      width = width,
-      height = height,
-      units = "cm",
-      res = plotConfiguration$res,
-      pointsize = plotConfiguration$pointsize
-    )
-  }
+  defaultPlotConfiguration$titleSize <- 8
+  defaultPlotConfiguration$xLabelSize <- 8
+  defaultPlotConfiguration$yLabelSize <- 8
+  defaultPlotConfiguration$xAxisLabelTicksSize <- 8
+  defaultPlotConfiguration$yAxisLabelTicksSize <- 8
+  defaultPlotConfiguration$legendTitleSize <- 6
+  defaultPlotConfiguration$legendPosition <- tlf::LegendPositions$outsideTop
+
+  return(defaultPlotConfiguration)
 }
 
-#' Close output device
+#' @title Create an instance of `PlotGridConfiguration` R6 class
+#' @rdname createEsqlabsPlotGridConfiguration
 #'
-#' @param plotConfiguration An object of type `PlotConfiguration`
-#' @import ospsuite
+#' @description
 #'
-#' @details If the output of the plot is directed to a file, close the device.
-#' @keywords internal
-closeOutputDevice <- function(plotConfiguration) {
-  if (is.null(plotConfiguration$outputDevice)) {
-    return()
-  }
-  if (enumHasKey(plotConfiguration$outputDevice, GraphicsDevices)) {
-    dev.off()
-  }
+#' An instance of `PlotGridConfiguration` R6 class from `{tlf}` package is
+#' needed for creating a grid of multiple visualizations created using the
+#' `{ospsuite}` package.
+#'
+#' The default attributes of the class are chosen to reflect the corporate
+#' standards adopted by esqLABS GmbH.
+#'
+#' @return An instance of `PlotGridConfiguration` R6 class.
+#'
+#' @examples
+#' createEsqlabsPlotGridConfiguration()
+#'
+#' @family create-plotting-configurations
+#'
+#' @export
+createEsqlabsPlotGridConfiguration <- function() {
+  plotGridConfiguration <- tlf::PlotGridConfiguration$new()
+
+  plotGridConfiguration$tagLevels <- "a"
+
+  return(plotGridConfiguration)
 }
 
-#' Does the plot type contain points?
+#' @param projectConfiguration Object of class `ProjectConfiguration`
+#' that contains information about the output paths.
 #'
-#' @param type String value of argument `type` passed to function `plot()`
+#' @title Create an instance of `ExportConfiguration` R6 class
+#' @rdname createEsqlabsExportConfiguration
 #'
-#' @return TRUE if `type` contains "p" or is "b", FALSE otherwise
+#' @description
 #'
-#' @keywords internal
-.isPoint <- function(type) {
-  isCharInString("p", type) || (type == "b")
-}
+#' An instance of `ExportConfiguration` R6 class from `{tlf}` package is needed
+#' for saving the plots and plot grids created using the `{ospsuite}` package.
+#'
+#' The default attributes of the class are chosen to reflect the corporate
+#' standards adopted by esqLABS GmbH.
+#'
+#' @return An instance of `ExportConfiguration` R6 class.
+#'
+#' @examples
+#' myProjConfig <- ProjectConfiguration$new()
+#' createEsqlabsExportConfiguration(myProjConfig)
+#'
+#' @family create-plotting-configurations
+#'
+#' @export
+createEsqlabsExportConfiguration <- function(projectConfiguration) {
+  exportConfiguration <- tlf::ExportConfiguration$new()
 
-#' Does the plot type contain line?
-#'
-#' @param type String value of argument `type` passed to function `plot()`
-#'
-#' @return TRUE if `type` contains "l" or is "b", FALSE otherwise
-#'
-#' @keywords internal
-.isLine <- function(type) {
-  isCharInString("l", type) || (type == "b")
+  exportConfiguration$path <- projectConfiguration$outputFolder
+  exportConfiguration$dpi <- 300
+  exportConfiguration$format <- projectConfiguration$outputDevice
+
+  exportConfiguration$width <- 18
+  exportConfiguration$height <- 18
+  exportConfiguration$units <- "cm"
+
+  return(exportConfiguration)
 }