+++ R/utilities.R
@@ -76,41 +76,6 @@ geosd <- function(x, na.rm = FALSE) {
   exp(sd(log(x), na.rm = na.rm))
 }
 
-#' Calculate quantiles for given xy-vectors
-#'
-#' @param quantiles A numerical vector with quantile values. Default is
-#' `c(0.05, 0.5, 0.95)`
-#' @param xValues X values, by which y values are grouped
-#' @param yValues Values for which the quantiles are calculated
-#'
-#' @return A list with `xValues` and aggregated `yValues`
-#' @export
-getQuantilesYData <- function(xValues, yValues, quantiles = c(0.05, 0.5, 0.95)) {
-  validateIsNumeric(c(xValues, yValues, quantiles))
-  validateIsSameLength(xValues, yValues)
-  output <- list()
-  # Aggregate time values
-  for (quantile in quantiles) {
-    aggregatedData <- aggregate(yValues, by = list(xVals = xValues), FUN = quantile, quantile)
-    output[[as.character(quantile)]] <- list()
-    output[[as.character(quantile)]][["xValues"]] <- aggregatedData$xVals
-    output[[as.character(quantile)]][["yValues"]] <- aggregatedData[[2]]
-  }
-
-  return(output)
-}
-
-#' Get hash code of the .NET object
-#'
-#' @param netWrapper Any object from the ospsuite-R that inherits from DotNetWrapper
-#' @import rClr
-#'
-#' @return Value of the .NET-method "GetHashCode"
-getNetHashCode <- function(netWrapper) {
-  validateIsOfType(netWrapper, "DotNetWrapper")
-  rClr::clrGet(netWrapper$ref, "HashCode")
-}
-
 
 #' Remove an entry from a list
 #'
@@ -138,35 +103,20 @@ removeFromList <- function(entry, listArg) {
   return(listArg)
 }
 
-#' Compare values including NA
+#' Compare values including `NA`
 #'
-#' @param v1 Value or a list of values to compare. May include NA.
-#' @param v2 Value or a list of values to compare. May include NA.
+#' @param v1 Value or a list of values to compare. May include `NA`.
+#' @param v2 Value or a list of values to compare. May include `NA`.
 #' @details From http://www.cookbook-r.com/Manipulating_data/Comparing_vectors_or_factors_with_NA/
 #'
-#' @return TRUE wherever elements are the same, including NA's,
-# and FALSE everywhere else.
+#' @return `TRUE` wherever elements are the same, including `NA`'s,
+# and `FALSE` everywhere else.
 #' @export
 compareWithNA <- function(v1, v2) {
   same <- (v1 == v2) | (is.na(v1) & is.na(v2))
   same[is.na(same)] <- FALSE
   return(same)
 }
-
-#' Is a character part of string?
-#'
-#' @param char Character to find in the string
-#' @param string String that should contain the character
-#'
-#' @return TRUE if the `character` is a substring if `string`, FALSE otherwise
-#' @export
-#'
-#' @examples
-#' isCharInString("a", "bsdalk")
-isCharInString <- function(char, string) {
-  any(unlist(strsplit(string, ""), use.names = FALSE) == char)
-}
-
 
 #' Escape a string for possible regular expression match
 #'