+++ R/zzz.R
@@ -1,4 +1,6 @@
+# nocov start
 .onLoad <- function(...) {
   options(warnPartialMatchDollar = TRUE)
   Sys.setenv("_R_CHECK_LENGTH_1_CONDITION_" = "true")
 }
+# nocov end